Move toast context to root of react folder
